Require original author credentials to delete blog, extract users into into request.user using custom middleware

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
@@ -29,16 +28,11 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   try {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!request.token || !decodedToken.id) {
-      return response
-        .status(401)
-        .json({ error: 'token missing or invalid token' })
-    }
-    const user = await User.findById(decodedToken.id)
+    // thanks to custom userExtractor middleware
+    const user = request.user
     const blog = new Blog({
       title: body.title,
       author: body.author,
@@ -83,13 +77,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (exception) {
-    next(exception)
-  }
-})
+blogsRouter.delete(
+  '/:id',
+  userExtractor,
+  async (request, response, next) => {
+    try {
+      const user = request.user
+      const blog = await Blog.findById(request.params.id)
+      // with userExtractor middleware instead of using !== decodedToken.id
+      if (blog.user.toString() !== user._id.toString()) {
+        return response.status(401).json({
+          error: 'only the original blog poster can delete this blog',
+        })
+      }
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } catch (exception) {
+      next(exception)
+    }
+  },
+)
 
 module.exports = blogsRouter
